requirements.txt
selenium>=4.0.0
webdriver-manager>=3.8.0
python-dotenv>=0.20.0  # To load environment variables from a .env file locally